+++ explainerdashboard/__init__.py
@@ -1,5 +1,2 @@
-# import dashboard_tabs
-# import dashboards
-# import explainers
-# import explainer_methods
-# import explainer_plots
+from .explainers import ClassifierExplainer, RegressionExplainer
+from .dashboards import ExplainerDashboard, InlineExplainer